// smartprecoapp/api/model/productDto.ts
/**
 * Generated by orval v7.9.0 🍺
 * Do not edit manually.
 * SmartPreço API
 * SmartPreço API é uma API RESTful que fornece informações sobre mercados, produtos e seus preços.
 * OpenAPI spec version: 1.0
 */

export interface ProductDto {
  /** Update timestamp */
  updatedAt: string;
  /** Product's image URL */
  imageUrl?: string;

  image_url?: string;
  /** Unique identifier of the product */
  id: string;
  /** Name of the product */
  name: string;
  /** Detailed description of the product */
  description: string;
  /** Category to which the product belongs */
  category: string;
  /** Lowest available moderated price for the product */
  lowestPrice?: number;
}